Carve `TestReductions` out of the monolith (wala/ML#635).
The 40 reduction and rounding tests (`reduce_*`, `argmax`/`argmin`,
`top_k`, `ceil`) move verbatim; the monolith keeps 790, preserving the
1,040 total.
